Don't be explicit about the address being a /32 or /128
Apparently being explicit about the CIDR suffix causes issues on
Chromebook and Android. It is very strange though because those
are the defaults anyway.

This is according to this thread and a user's comment on how
they fixed a problem they faced with connectivity
https://github.com/rany2/warp.sh/issues/6

I don't have such a problem myself but seeing that this is
just relying on the implicit behavior it should be alright.

Signed-off-by: rany <ranygh@riseup.net>